LATEST VERSION with selectable Baud-rate and foreground/background Color selection for terminal text.
Latest Version with selectable Baud-rate and foreground/background Color selection for terminal text.
The corresponding new options have been added to the Settings screen, accesible with Ctrl-F10.
When a color change is detected, exiting the setup screen, a reboot of the Propeller if performed. This is transparent to the connected Host (like an RC2014 system).